feat: simplify qualification board
Remove redundant columns elements from qualification board, and changed them to simple grids.
Adjust the row height.
