--- frontend_scoring/src/features/ScoreBoard/QualifyingPhaseBoard/ColumnElements/InstitutionColumnElement.tsx
@@ -1,9 +1,0 @@
-interface Props {
-  content: string;
-}
-
-function InstitutionsColumnElement(props: Props) {
-  return <div className="institution_column_element">{props.content}</div>;
-}
-
-export default InstitutionsColumnElement;
