--- frontend_scoring/src/features/ScoreBoard/QualifyingPhaseBoard/ColumnElements/NameColumnElement.tsx
@@ -1,9 +1,0 @@
-interface Props {
-  content: string;
-}
-
-function NameColumnElement(props: Props) {
-  return <div className="name_column_element">{props.content}</div>;
-}
-
-export default NameColumnElement;
