--- frontend_scoring/src/features/ScoreBoard/QualifyingPhaseBoard/ColumnElements/RankColumnElement.tsx
@@ -1,9 +1,0 @@
-interface Props {
-  content: string;
-}
-
-function RankColumnElement(props: Props) {
-  return <div className="rank_column_element">{props.content}</div>;
-}
-
-export default RankColumnElement;
