--- frontend_scoring/src/features/ScoreBoard/QualifyingPhaseBoard/ColumnElements/ScoreColumnElement.tsx
@@ -1,9 +1,0 @@
-interface Props {
-  content: string;
-}
-
-function ScoreColumnElement(props: Props) {
-  return <div className="score_column_element">{props.content}</div>;
-}
-
-export default ScoreColumnElement;
