--- frontend_scoring/src/features/ScoreBoard/QualifyingPhaseBoard/ColumnElements/TargetColumnElement.tsx
@@ -1,9 +1,0 @@
-interface Props {
-  content: string;
-}
-
-function TargetColumnElement(props: Props) {
-  return <div className="target_column_element">{props.content}</div>;
-}
-
-export default TargetColumnElement;
